requirements.txt
google-generativeai
python-dotenv
fastapi
uvicorn[standard]
pydantic
python-multipart
